+++ crates/cheatnet/tests/cheatcodes/roll.rs
@@ -9,6 +9,7 @@ use crate::{
 use cairo_felt::Felt252;
 use cheatnet::cheatcodes::deploy::deploy;
 use cheatnet::rpc::call_contract;
+use cheatnet::state::CheatTarget;
 use conversions::StarknetConversions;
 
 #[test]
@@ -23,7 +24,7 @@ fn roll_simple() {
         &[],
     );
 
-    cheatnet_state.start_roll(contract_address, Felt252::from(123_u128));
+    cheatnet_state.start_roll(CheatTarget::One(contract_address), Felt252::from(123_u128));
 
     let selector = felt_selector_from_name("get_block_number");
 
@@ -51,7 +52,7 @@ fn roll_with_other_syscall() {
         &[],
     );
 
-    cheatnet_state.start_roll(contract_address, Felt252::from(123_u128));
+    cheatnet_state.start_roll(CheatTarget::One(contract_address), Felt252::from(123_u128));
 
     let selector = felt_selector_from_name("get_block_number_and_emit_event");
 
@@ -80,7 +81,10 @@ fn roll_in_constructor() {
         .unwrap();
     let precalculated_address = cheatnet_state.precalculate_address(&class_hash, &[]);
 
-    cheatnet_state.start_roll(precalculated_address, Felt252::from(123_u128));
+    cheatnet_state.start_roll(
+        CheatTarget::One(precalculated_address),
+        Felt252::from(123_u128),
+    );
 
     let contract_address = deploy(&mut blockifier_state, &mut cheatnet_state, &class_hash, &[])
         .unwrap()
@@ -127,7 +131,7 @@ fn roll_stop() {
 
     let old_block_number = recover_data(output);
 
-    cheatnet_state.start_roll(contract_address, Felt252::from(123_u128));
+    cheatnet_state.start_roll(CheatTarget::One(contract_address), Felt252::from(123_u128));
 
     let output = call_contract(
         &mut blockifier_state,
@@ -142,7 +146,7 @@ fn roll_stop() {
     assert_eq!(new_block_number, vec![Felt252::from(123)]);
     assert_ne!(old_block_number, new_block_number);
 
-    cheatnet_state.stop_roll(contract_address);
+    cheatnet_state.stop_roll(CheatTarget::One(contract_address));
 
     let output = call_contract(
         &mut blockifier_state,
@@ -182,8 +186,8 @@ fn roll_double() {
 
     let old_block_number = recover_data(output);
 
-    cheatnet_state.start_roll(contract_address, Felt252::from(123_u128));
-    cheatnet_state.start_roll(contract_address, Felt252::from(123_u128));
+    cheatnet_state.start_roll(CheatTarget::One(contract_address), Felt252::from(123_u128));
+    cheatnet_state.start_roll(CheatTarget::One(contract_address), Felt252::from(123_u128));
 
     let output = call_contract(
         &mut blockifier_state,
@@ -198,7 +202,7 @@ fn roll_double() {
     assert_eq!(new_block_number, vec![Felt252::from(123)]);
     assert_ne!(old_block_number, new_block_number);
 
-    cheatnet_state.stop_roll(contract_address);
+    cheatnet_state.stop_roll(CheatTarget::One(contract_address));
 
     let output = call_contract(
         &mut blockifier_state,
@@ -242,7 +246,7 @@ fn roll_proxy() {
     )
     .unwrap();
 
-    cheatnet_state.start_roll(contract_address, Felt252::from(123_u128));
+    cheatnet_state.start_roll(CheatTarget::One(contract_address), Felt252::from(123_u128));
 
     let after_roll_output = call_contract(
         &mut blockifier_state,
@@ -255,7 +259,7 @@ fn roll_proxy() {
 
     assert_success!(after_roll_output, vec![Felt252::from(123)]);
 
-    cheatnet_state.stop_roll(contract_address);
+    cheatnet_state.stop_roll(CheatTarget::One(contract_address));
 
     let after_roll_cancellation_output = call_contract(
         &mut blockifier_state,
@@ -297,7 +301,7 @@ fn roll_library_call() {
     )
     .unwrap();
 
-    cheatnet_state.start_roll(lib_call_address, Felt252::from(123_u128));
+    cheatnet_state.start_roll(CheatTarget::One(lib_call_address), Felt252::from(123_u128));
 
     let after_roll_output = call_contract(
         &mut blockifier_state,
@@ -310,7 +314,7 @@ fn roll_library_call() {
 
     assert_success!(after_roll_output, vec![Felt252::from(123)]);
 
-    cheatnet_state.stop_roll(lib_call_address);
+    cheatnet_state.stop_roll(CheatTarget::One(lib_call_address));
 
     let after_roll_cancellation_output = call_contract(
         &mut blockifier_state,
@@ -322,4 +326,247 @@ fn roll_library_call() {
     .unwrap();
 
     assert_outputs(before_roll_output, after_roll_cancellation_output);
+}
+
+#[test]
+fn roll_all_simple() {
+    let mut cached_state = create_cached_state();
+    let (mut blockifier_state, mut cheatnet_state) = create_cheatnet_state(&mut cached_state);
+
+    let contract_address = deploy_contract(
+        &mut blockifier_state,
+        &mut cheatnet_state,
+        "RollChecker",
+        &[],
+    );
+
+    cheatnet_state.start_roll(CheatTarget::All, Felt252::from(123));
+
+    let selector = felt_selector_from_name("get_block_number");
+
+    let output = call_contract(
+        &mut blockifier_state,
+        &mut cheatnet_state,
+        &contract_address,
+        &selector,
+        &[],
+    )
+    .unwrap();
+
+    assert_success!(output, vec![Felt252::from(123)]);
+}
+
+#[test]
+fn roll_all_then_one() {
+    let mut cached_state = create_cached_state();
+    let (mut blockifier_state, mut cheatnet_state) = create_cheatnet_state(&mut cached_state);
+
+    let contract_address = deploy_contract(
+        &mut blockifier_state,
+        &mut cheatnet_state,
+        "RollChecker",
+        &[],
+    );
+
+    cheatnet_state.start_roll(CheatTarget::All, Felt252::from(321_u128));
+    cheatnet_state.start_roll(CheatTarget::One(contract_address), Felt252::from(123_u128));
+
+    let selector = felt_selector_from_name("get_block_number");
+
+    let output = call_contract(
+        &mut blockifier_state,
+        &mut cheatnet_state,
+        &contract_address,
+        &selector,
+        &[],
+    )
+    .unwrap();
+
+    assert_success!(output, vec![Felt252::from(123)]);
+}
+
+#[test]
+fn roll_one_then_all() {
+    let mut cached_state = create_cached_state();
+    let (mut blockifier_state, mut cheatnet_state) = create_cheatnet_state(&mut cached_state);
+
+    let contract_address = deploy_contract(
+        &mut blockifier_state,
+        &mut cheatnet_state,
+        "RollChecker",
+        &[],
+    );
+
+    cheatnet_state.start_roll(CheatTarget::One(contract_address), Felt252::from(123_u128));
+    cheatnet_state.start_roll(CheatTarget::All, Felt252::from(321_u128));
+
+    let selector = felt_selector_from_name("get_block_number");
+
+    let output = call_contract(
+        &mut blockifier_state,
+        &mut cheatnet_state,
+        &contract_address,
+        &selector,
+        &[],
+    )
+    .unwrap();
+
+    assert_success!(output, vec![Felt252::from(321)]);
+}
+
+#[test]
+fn roll_all_stop() {
+    let mut cached_state = create_cached_state();
+    let (mut blockifier_state, mut cheatnet_state) = create_cheatnet_state(&mut cached_state);
+
+    let contract_address = deploy_contract(
+        &mut blockifier_state,
+        &mut cheatnet_state,
+        "RollChecker",
+        &[],
+    );
+
+    let selector = felt_selector_from_name("get_block_number");
+
+    let output = call_contract(
+        &mut blockifier_state,
+        &mut cheatnet_state,
+        &contract_address,
+        &selector,
+        &[],
+    )
+    .unwrap();
+
+    let old_block_number = recover_data(output);
+
+    cheatnet_state.start_roll(CheatTarget::All, Felt252::from(123_u128));
+
+    let output = call_contract(
+        &mut blockifier_state,
+        &mut cheatnet_state,
+        &contract_address,
+        &selector,
+        &[],
+    )
+    .unwrap();
+
+    let new_block_number = recover_data(output);
+    assert_eq!(new_block_number, vec![Felt252::from(123)]);
+    assert_ne!(old_block_number, new_block_number);
+
+    cheatnet_state.stop_roll(CheatTarget::All);
+
+    let output = call_contract(
+        &mut blockifier_state,
+        &mut cheatnet_state,
+        &contract_address,
+        &selector,
+        &[],
+    )
+    .unwrap();
+    let changed_back_block_number = recover_data(output);
+
+    assert_eq!(old_block_number, changed_back_block_number);
+}
+
+#[test]
+fn roll_multiple() {
+    let mut cached_state = create_cached_state();
+    let (mut blockifier_state, mut cheatnet_state) = create_cheatnet_state(&mut cached_state);
+
+    let contract = "RollChecker".to_owned().to_felt252();
+    let contracts = get_contracts();
+    let class_hash = blockifier_state.declare(&contract, &contracts).unwrap();
+
+    let contract_address1 = deploy(&mut blockifier_state, &mut cheatnet_state, &class_hash, &[])
+        .unwrap()
+        .contract_address;
+
+    let contract_address2 = deploy(&mut blockifier_state, &mut cheatnet_state, &class_hash, &[])
+        .unwrap()
+        .contract_address;
+
+    let selector = felt_selector_from_name("get_block_number");
+
+    let output = call_contract(
+        &mut blockifier_state,
+        &mut cheatnet_state,
+        &contract_address1,
+        &selector,
+        &[],
+    )
+    .unwrap();
+
+    let old_block_number1 = recover_data(output);
+
+    let output = call_contract(
+        &mut blockifier_state,
+        &mut cheatnet_state,
+        &contract_address2,
+        &selector,
+        &[],
+    )
+    .unwrap();
+
+    let old_block_number2 = recover_data(output);
+
+    cheatnet_state.start_roll(
+        CheatTarget::Multiple(vec![contract_address1, contract_address2]),
+        Felt252::from(123_u128),
+    );
+
+    let output = call_contract(
+        &mut blockifier_state,
+        &mut cheatnet_state,
+        &contract_address1,
+        &selector,
+        &[],
+    )
+    .unwrap();
+
+    let new_block_number1 = recover_data(output);
+
+    let output = call_contract(
+        &mut blockifier_state,
+        &mut cheatnet_state,
+        &contract_address2,
+        &selector,
+        &[],
+    )
+    .unwrap();
+
+    let new_block_number2 = recover_data(output);
+
+    assert_eq!(new_block_number1, vec![Felt252::from(123)]);
+    assert_eq!(new_block_number2, vec![Felt252::from(123)]);
+
+    cheatnet_state.stop_roll(CheatTarget::Multiple(vec![
+        contract_address1,
+        contract_address2,
+    ]));
+
+    let output = call_contract(
+        &mut blockifier_state,
+        &mut cheatnet_state,
+        &contract_address1,
+        &selector,
+        &[],
+    )
+    .unwrap();
+
+    let changed_back_block_number1 = recover_data(output);
+
+    let output = call_contract(
+        &mut blockifier_state,
+        &mut cheatnet_state,
+        &contract_address2,
+        &selector,
+        &[],
+    )
+    .unwrap();
+
+    let changed_back_block_number2 = recover_data(output);
+
+    assert_eq!(old_block_number1, changed_back_block_number1);
+    assert_eq!(old_block_number2, changed_back_block_number2);
 }